Cron/ncaamb/requirements.txt
catboost==1.2.8
lightgbm==4.6.0
matplotlib==3.10.8
mysql_connector_repackaged==0.3.1
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
polars==1.36.1
python-dotenv==1.2.1
pytz==2025.1
Requests==2.32.5
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
SQLAlchemy==2.0.45
undetected_chromedriver==3.5.5
xgboost==3.1.2